Add files via upload
The code is giving the solution to weighted Linear Regression equation. It is coded in R-Studios, however we use C++ using the RcppEigen plug-in. We then develop a function in R using its custom functions. We compare the time using microbenchmark package. In C++, we have used Eigen Library to do matrix multiplications and transformations.