Add files via upload
The algorithm generates solution of a Weighted Linear Regression equation. This code uses the Rcpp package in R, to import C++ code. We don't use Eigen in C++ code. C++ is slower in matrix operations as compared to the inbuilt R functions. However, the story is different when we use Eigen library, which is done in the second code.