Add files via upload
In this code, we focus on basic stock analysis. We plot the prices of two stocks, check for correlation in price movement. We also visualize prices and returns. For plots we use Autoplot, a part of ggplot2 which works on time series objects. We then compare the process for different stocks.